Add a new TestExtensionService for testing (needed to validate sort order)
This change also removes the copy-pasta in TestExtensionService*
