+++ Assets/MixedRealityToolkit.Tests/EditModeTests/Services/BaseTestExtensionService.cs
@@ -1,0 +1,34 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+
+namespace Microsoft.MixedReality.Toolkit.Tests.Services
+{
+    /// <summary>
+    /// A base class for test extension services.
+    /// </summary>
+    internal abstract class BaseTestExtensionService : BaseExtensionService
+    {
+        public BaseTestExtensionService(
+            IMixedRealityServiceRegistrar registrar,
+            string name,
+            uint priority,
+            BaseMixedRealityProfile profile) : base(registrar, name, priority, profile) { }
+
+        public bool IsEnabled { get; private set; }
+
+        public override void Enable()
+        {
+            IsEnabled = true;
+        }
+
+        public override void Disable()
+        {
+            IsEnabled = false;
+        }
+
+        public override void Destroy()
+        {
+        }
+    }
+}
