+++ Assets/MixedRealityToolkit.Tests/EditModeTests/Services/BaseTestExtensionService.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7600bde8a525f664fbcda11fe0d23104
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
