+++ Assets/MixedRealityToolkit.Tests/EditModeTests/Services/ITestExtensionService3.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b6d95e2f35b94af458f51b7244ecc584
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
