+++ Assets/MixedRealityToolkit.Tests/EditModeTests/Services/TestExtensionService2.cs
@@ -4,28 +4,12 @@
 
 namespace Microsoft.MixedReality.Toolkit.Tests.Services
 {
-    internal class TestExtensionService2 : BaseExtensionService, ITestExtensionService2
+    internal class TestExtensionService2 : BaseTestExtensionService, ITestExtensionService2
     {
         public TestExtensionService2(
             IMixedRealityServiceRegistrar registrar, 
             string name, 
             uint priority, 
             BaseMixedRealityProfile profile) : base(registrar, name, priority, profile) { }
-
-        public bool IsEnabled { get; private set; }
-
-        public override void Enable()
-        {
-            IsEnabled = true;
-        }
-
-        public override void Disable()
-        {
-            IsEnabled = false;
-        }
-
-        public override void Destroy()
-        {
-        }
     }
 }
