+++ Assets/MixedRealityToolkit.Tests/EditModeTests/Services/TestExtensionService3.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9538552f6f452364e91f213efa5b8f36
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
